moving to pixel int

+++ python_analytics/csknow-python-analytics/visibility/recordGameplay.py
@@ -76,7 +76,7 @@ for config in args.configs:
         print("couldn't find demo playback name")
         quit(1)
     pydirectinput.moveTo(demo_playback_name_location.left + demo_playback_name_location.width,
-                         demo_playback_name_location.top + demo_playback_name_location.height / 2)
+                         int(demo_playback_name_location.top + demo_playback_name_location.height / 2))
     pydirectinput.mouseDown(button='left')
     time.sleep(1.0)
     pydirectinput.moveTo(-215, -324)
@@ -92,8 +92,8 @@ for config in args.configs:
     if demo_playback_name_location is None:
         print("couldn't find console text entry")
         quit(1)
-    pydirectinput.moveTo(console_text_entry_location.left + console_text_entry_location.width / 3,
-                         console_text_entry_location.top + console_text_entry_location.height / 2)
+    pydirectinput.moveTo(int(console_text_entry_location.left + console_text_entry_location.width / 3),
+                         int(console_text_entry_location.top + console_text_entry_location.height / 2))
     pydirectinput.click()
     #pyautogui.write('mirv_streams previewEnd\n')
     #time.sleep(0.5)
@@ -117,8 +117,8 @@ for config in args.configs:
         if location is None:
             found_end = False
 
-    pydirectinput.moveTo(console_text_entry_location.left + console_text_entry_location.width / 3,
-                         console_text_entry_location.top + console_text_entry_location.height / 2)
+    pydirectinput.moveTo(int(console_text_entry_location.left + console_text_entry_location.width / 3),
+                         int(console_text_entry_location.top + console_text_entry_location.height / 2))
     pydirectinput.click()
     pydirectinput.press('F2')
     pydirectinput.press('`')
